logos: move all SVGs to one location for Component Import
